fix: widen the elected subproblem count to BIGINT
An audit of the previous commit found the (int) cast on the return is signed
overflow for M above ~6.4e13, which MAX_NU_PTS=1e14 permits: nb can exceed
INT_MAX when the occupancy is low and nthreads small. Return BIGINT and store
BIGINT in the plan.

Dropping the npts clamp with it, so the elected count is unconditionally a
whole multiple of nthreads - the caller already clips to M, and the clamp was
the only case that broke the invariant the docstring states (and that the
property test asserts). Test grid gains npts=10, which on 128 threads is the
npts < nthreads corner that had no coverage.

Election-identical at the caller across 577668 configs (dim 2-3, ns 3-16,
integral npts 2..1e9, 13 thread counts, occupancy 1..npts). 27/27.

Also drops the spread_thread=2 caveat from the call site, where it duplicated
the docstring, and trims the decline evidence to the claim plus its size.

+++ include/finufft/spreadinterp.hpp
@@ -206,9 +206,9 @@ void FINUFFT_PLAN_T<TF>::indexSort()
                                                         (double)m.nOccupiedBins, bin_size,
                                                         m.spopts.nthreads);
     if (m.spopts.debug && m.nSubproblems)
-      printf("\toccupancy %.3g pts/bin -> nb=%d (%.3g pts/subproblem)\n",
-             m.nOccupiedBins ? (double)M / (double)m.nOccupiedBins : 0.0, m.nSubproblems,
-             (double)M / (double)m.nSubproblems);
+      printf("\toccupancy %.3g pts/bin -> nb=%lld (%.3g pts/subproblem)\n",
+             m.nOccupiedBins ? (double)M / (double)m.nOccupiedBins : 0.0,
+             (long long)m.nSubproblems, (double)M / (double)m.nSubproblems);
   }
 }
